requirements.txt
fastapi>=0.104.0
hypercorn==0.14.4
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
pydantic-settings>=2.0.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.28.0
alembic>=1.12.0

# Redis and Queue
redis>=5.0.0
rq>=1.15.0
rq-dashboard>=0.7.0

# HTTP Client
httpx>=0.25.0
aiohttp>=3.9.0

# External APIs
openai>=1.0.0
apify-client>=1.6.0
google-api-python-client>=2.0.0

# Salesforce
simple-salesforce>=1.12.0
requests-oauthlib>=1.3.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
email-validator>=2.0.0
phonenumbers>=8.13.0

# Monitoring and Logging
structlog>=23.0.0
prometheus-client>=0.17.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Security
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
